Initial commit on the F_separate_lib_CLI branch
Started work to separate the project into a Rust library and a separate
CLI.

The library in lib/ provides the main syswall functionality which is
exposed via the syswall::trace() function.

The CLI application in cli/ uses the syswall library as a dependency and
performs tracing as before using syswall::trace.  It also handles
command-line argument parsing, saving/loading of configuration, etc.

A new RuntimeConf struct has been added which contains a syscall_cb
item, which is a function that is called whenever a user decision is
required.  This function should retrieve the response and return it as a
UserResponse.  In the CLI, this is handled by asking the user to enter
the command via stdin, as before.  This separation however allows
clients to respond to decisions in any appropriate way based on the type
of client.